requirements.txt
async-timeout==4.0.3
blinker==1.7.0
click==8.1.7
Flask==3.0.0
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
python-dotenv==1.0.0
redis==5.0.1
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
